Making Authorized Requests

+++ frontend/src/components/WorkoutDetails.js
@@ -1,11 +1,19 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext.js";
+import { useAuthContext } from "../hooks/useAuthContext.js";
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
 
   const handleClick = async () => {
+    if (!user) {
+      return;
+    }
     const response = await fetch(`/api/workouts/${workout._id}`, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
     const data = await response.json();
 
@@ -28,7 +36,9 @@ const WorkoutDetails = ({ workout }) => {
         {workout.reps}
       </p>
       <p>{workout.createdAt}</p>
-      <span onClick={handleClick}>delete</span>
+      <span className="material-symbols-outlined" onClick={handleClick}>
+        delete
+      </span>
     </div>
   );
 };
